Split out some content from types.go
No functional changes
