+++ errors.go
@@ -1,0 +1,112 @@
+package amqp
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/Azure/go-amqp/internal/buffer"
+)
+
+// ErrorCondition is one of the error conditions defined in the AMQP spec.
+type ErrorCondition string
+
+func (ec ErrorCondition) marshal(wr *buffer.Buffer) error {
+	return (symbol)(ec).marshal(wr)
+}
+
+func (ec *ErrorCondition) unmarshal(r *buffer.Buffer) error {
+	s, err := readString(r)
+	*ec = ErrorCondition(s)
+	return err
+}
+
+// Error Conditions
+const (
+	// AMQP Errors
+	ErrorInternalError         ErrorCondition = "amqp:internal-error"
+	ErrorNotFound              ErrorCondition = "amqp:not-found"
+	ErrorUnauthorizedAccess    ErrorCondition = "amqp:unauthorized-access"
+	ErrorDecodeError           ErrorCondition = "amqp:decode-error"
+	ErrorResourceLimitExceeded ErrorCondition = "amqp:resource-limit-exceeded"
+	ErrorNotAllowed            ErrorCondition = "amqp:not-allowed"
+	ErrorInvalidField          ErrorCondition = "amqp:invalid-field"
+	ErrorNotImplemented        ErrorCondition = "amqp:not-implemented"
+	ErrorResourceLocked        ErrorCondition = "amqp:resource-locked"
+	ErrorPreconditionFailed    ErrorCondition = "amqp:precondition-failed"
+	ErrorResourceDeleted       ErrorCondition = "amqp:resource-deleted"
+	ErrorIllegalState          ErrorCondition = "amqp:illegal-state"
+	ErrorFrameSizeTooSmall     ErrorCondition = "amqp:frame-size-too-small"
+
+	// Connection Errors
+	ErrorConnectionForced   ErrorCondition = "amqp:connection:forced"
+	ErrorFramingError       ErrorCondition = "amqp:connection:framing-error"
+	ErrorConnectionRedirect ErrorCondition = "amqp:connection:redirect"
+
+	// Session Errors
+	ErrorWindowViolation  ErrorCondition = "amqp:session:window-violation"
+	ErrorErrantLink       ErrorCondition = "amqp:session:errant-link"
+	ErrorHandleInUse      ErrorCondition = "amqp:session:handle-in-use"
+	ErrorUnattachedHandle ErrorCondition = "amqp:session:unattached-handle"
+
+	// Link Errors
+	ErrorDetachForced          ErrorCondition = "amqp:link:detach-forced"
+	ErrorTransferLimitExceeded ErrorCondition = "amqp:link:transfer-limit-exceeded"
+	ErrorMessageSizeExceeded   ErrorCondition = "amqp:link:message-size-exceeded"
+	ErrorLinkRedirect          ErrorCondition = "amqp:link:redirect"
+	ErrorStolen                ErrorCondition = "amqp:link:stolen"
+)
+
+/*
+<type name="error" class="composite" source="list">
+    <descriptor name="amqp:error:list" code="0x00000000:0x0000001d"/>
+    <field name="condition" type="symbol" requires="error-condition" mandatory="true"/>
+    <field name="description" type="string"/>
+    <field name="info" type="fields"/>
+</type>
+*/
+
+// Error is an AMQP error.
+type Error struct {
+	// A symbolic value indicating the error condition.
+	Condition ErrorCondition
+
+	// descriptive text about the error condition
+	//
+	// This text supplies any supplementary details not indicated by the condition field.
+	// This text can be logged as an aid to resolving issues.
+	Description string
+
+	// map carrying information about the error condition
+	Info map[string]interface{}
+}
+
+func (e *Error) marshal(wr *buffer.Buffer) error {
+	return marshalComposite(wr, typeCodeError, []marshalField{
+		{value: &e.Condition, omit: false},
+		{value: &e.Description, omit: e.Description == ""},
+		{value: e.Info, omit: len(e.Info) == 0},
+	})
+}
+
+func (e *Error) unmarshal(r *buffer.Buffer) error {
+	return unmarshalComposite(r, typeCodeError, []unmarshalField{
+		{field: &e.Condition, handleNull: func() error { return errors.New("Error.Condition is required") }},
+		{field: &e.Description},
+		{field: &e.Info},
+	}...)
+}
+
+func (e *Error) String() string {
+	if e == nil {
+		return "*Error(nil)"
+	}
+	return fmt.Sprintf("*Error{Condition: %s, Description: %s, Info: %v}",
+		e.Condition,
+		e.Description,
+		e.Info,
+	)
+}
+
+func (e *Error) Error() string {
+	return e.String()
+}
